// out/model/login.ts
/**
 * TransferTo Calculator API
 * TransferTo Calculator
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Login {
    /**
     * The email used for authentication
     */
    email?: string;
    /**
     * The password used for authentication
     */
    password?: string;
}