// out/model/userRead.ts
/**
 * TransferTo Calculator API
 * TransferTo Calculator
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { User } from './user';


export interface UserRead {
    /**
     * Middle name
     */
    middle_name?: string;
    /**
     * The email used for authentication
     */
    email?: string;
    /**
     * Last name
     */
    last_name?: string;
    /**
     * First name
     */
    first_name?: string;
    /**
     * The unique identifier of a user
     */
    id?: string;
    /**
     * Created at this timestamp
     */
    created_ts?: Date;
    /**
     * Deleted at this timestamp
     */
    deleted_ts?: Date;
    /**
     * Updated at this timestamp
     */
    last_modified_ts?: Date;
}